add thumbails preview
