+++ index.html
@@ -241,8 +241,7 @@
 			</div>
 		</section>
 	</template>
-	<!-- <script src="js/functions.js"></script>
-	<script src="js/test.ts"></script> -->
+	
 	<script src="js/main.js" type="module"></script>
 </body>
